+++ app.js
@@ -13,7 +13,7 @@ import mod2Object from './mod2Object';
 // * The star of the show * //
 // ex- new MilSym('.test', 'Infantry', 'friendly', 'team', 'Armored', 'Rail', 'A/2-101', '27/42ID', '+').placeSymbol();
 class MilSym {
-  constructor(location, symbol, affiliation = 'friendly', echelon = 'none', mod1 = 'None', mod2 = 'None', uniqueDesignation = '', higherFormation = '', reinforcedReduced = '', isFlying = false) {
+  constructor(location, symbol, affiliation = 'friendly', echelon = 'none', mod1 = 'None', mod2 = 'None', uniqueDesignation = '', higherFormation = '', reinforcedReduced = '', isFlying = false, isActivity = false) {
     this.location = document.querySelector(location);
     this.symbol = militarySymbolsObject[symbol].affiliation[affiliation];
     this.affiliation = affiliationOutlineObject[affiliation];
@@ -26,6 +26,7 @@ class MilSym {
     this.reinforcedReduced = reinforcedReduced;
     this.isFlying = isFlying;
     this.flightCapable = militarySymbolsObject[symbol].flightCapable;
+    this.isActivity = isActivity;
     this.data = {
       location,
       symbol,
@@ -56,8 +57,7 @@ class MilSym {
     // } else {
     //   svg.append(this.affiliationOutlineData, this.decoratorData, this.echelonData);
     // }
-
-    svg.append(this.affiliationOutlineData, this.decoratorData, this.echelonData, this.mod1Data, this.mod2Data, this.uniqueDesignationData, this.higherFormationData, this.reinforcedReducedData, this.isFlyingData);
+    svg.append(this.affiliationOutlineData, this.decoratorData, this.echelonData, this.mod1Data, this.mod2Data, this.uniqueDesignationData, this.higherFormationData, this.reinforcedReducedData, this.isFlyingData, this.activityData);
     this.location.append(svg);
     svg.setAttributeNS(null, 'data-symbol-name', this.data.symbol);
     svg.setAttributeNS(null, 'data-symbol-info', JSON.stringify(this.data)); // this should probably be split into separate data-attrs
@@ -424,6 +424,20 @@ class MilSym {
     reinforcedReducedText.textContent = this.reinforcedReduced;
     reinforcedReducedGroup.append(reinforcedReducedText);
     return reinforcedReducedGroup;
+  }
+
+  get activityData() {
+    if (this.isActivity) {
+      const activityGroup = document.createElementNS('http://www.w3.org/2000/svg', 'g');
+      activityGroup.classList.add('activity');
+      const activityModifier = document.createElementNS('http://www.w3.org/2000/svg', 'path');
+      activityModifier.setAttribute('d', this.affiliation.activity);
+      activityModifier.setAttribute('fill', 'black');
+      activityModifier.setAttribute('stroke', 'black');
+      activityModifier.setAttribute('stroke-width', '4');
+      activityGroup.append(activityModifier);
+      return activityGroup;
+    }
   }
 }